pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "aTrain"
description = "aTrain is a tool for automatically transcribing speech recordings utilizing state-of-the-art machine learning models without uploading any data."

authors = [
  { name = "Armin Haberl", email = "armin.haberl@uni-graz.at" },
  { name = "Jürgen Fleiß", email = "juergen.fleiss@uni-graz.at" },
  { name = "Dominik Kowald", email = "dkowald@know-center.at" },
  { name = "Stefan Thalmann", email = "stefan.thalmann@uni-graz.at" },
]

dynamic = ["version"]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"

dependencies = [
  "aTrain_core@git+https://github.com/JuergenFleiss/atrain_core.git@v1.3.1",
  "wakepy==0.7.2",
  "show-in-file-manager==1.1.4",
  "typer",
  "nicegui",
  "pywebview",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["aTrain*"]

[project.scripts]
aTrain = "aTrain.cli:cli"

[tool.setuptools.dynamic]
version = { attr = "aTrain.version.__version__" }
